# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sarkari Exams API",
    "version": "1.0.0",
    "description": "Shared API contract for frontend/backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/announcements/{slug}": {
      "get": {
        "summary": "Get announcement by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Announcement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Auth response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Auth response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookmarks": {
      "get": {
        "summary": "List bookmarks",
        "responses": {
          "200": {
            "description": "Bookmarks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add bookmark",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bookmarks/{id}": {
      "delete": {
        "summary": "Remove bookmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile": {
      "get": {
        "summary": "Get profile",
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/subscriptions": {
      "post": {
        "summary": "Create or update subscription",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContentType": {
        "type": "string",
        "enum": [
          "job",
          "result",
          "admit-card",
          "syllabus",
          "answer-key",
          "admission"
        ]
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "ImportantDate": {
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string"
          },
          "eventDate": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "eventName",
          "eventDate"
        ]
      },
      "Announcement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "category": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "externalLink": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          },
          "minQualification": {
            "type": "string"
          },
          "ageLimit": {
            "type": "string"
          },
          "applicationFee": {
            "type": "string"
          },
          "salaryMin": {
            "type": "number"
          },
          "salaryMax": {
            "type": "number"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "cutoffMarks": {
            "type": "string"
          },
          "totalPosts": {
            "type": "integer"
          },
          "postedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "viewCount": {
            "type": "integer"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "importantDates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportantDate"
            }
          },
          "jobDetails": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "type",
          "category",
          "organization",
          "postedAt",
          "isActive",
          "viewCount"
        ]
      },
      "AnnouncementResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Announcement"
          }
        },
        "required": [
          "data"
        ]
      },
      "AnnouncementListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Announcement"
            }
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              }
            },
            "required": [
              "user"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "challengeToken": {
            "type": "string"
          },
          "twoFactorCode": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "BookmarkCreateRequest": {
        "type": "object",
        "properties": {
          "announcementId": {
            "type": "string"
          }
        },
        "required": [
          "announcementId"
        ]
      },
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "SubscriptionRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "frequency": {
            "type": "string",
            "enum": [
              "instant",
              "daily",
              "weekly"
            ]
          }
        },
        "required": [
          "email"
        ]
      },
      "SubscriptionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "verified": {
                "type": "boolean"
              }
            },
            "required": [
              "verified"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "data"
        ]
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "preferredCategories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredQualifications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredLocations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredOrganizations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ageGroup": {
            "type": "string",
            "nullable": true
          },
          "educationLevel": {
            "type": "string",
            "nullable": true
          },
          "experienceYears": {
            "type": "integer"
          },
          "emailNotifications": {
            "type": "boolean"
          },
          "pushNotifications": {
            "type": "boolean"
          },
          "notificationFrequency": {
            "type": "string"
          },
          "profileComplete": {
            "type": "boolean"
          },
          "onboardingCompleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "userId",
          "preferredCategories",
          "preferredQualifications",
          "preferredLocations",
          "preferredOrganizations",
          "experienceYears",
          "emailNotifications",
          "pushNotifications",
          "notificationFrequency",
          "profileComplete",
          "onboardingCompleted"
        ]
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserProfile"
          }
        },
        "required": [
          "data"
        ]
      },
      "ProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "preferredCategories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredQualifications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredLocations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredOrganizations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ageGroup": {
            "type": "string",
            "nullable": true
          },
          "educationLevel": {
            "type": "string",
            "nullable": true
          },
          "experienceYears": {
            "type": "integer"
          },
          "emailNotifications": {
            "type": "boolean"
          },
          "pushNotifications": {
            "type": "boolean"
          },
          "notificationFrequency": {
            "type": "string"
          },
          "profileComplete": {
            "type": "boolean"
          },
          "onboardingCompleted": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}